norminette ok movement + get parsing infos
